move camera logic and change coordinates view to use cases
