+++ app/src/main/assets/paperwork.json
@@ -1,1 +1,1 @@
-{"buildTime":"2020-05-08 13:58","gitSha":"1f7cfeb95"}
+{"buildTime":"2020-05-11 16:25","gitSha":"e2c22663e"}
