+++ app/src/main/java/org/dhis2/data/forms/dataentry/tablefields/coordinate/CoordinateHolder.java
@@ -9,8 +9,8 @@ import android.widget.TextView;
 import org.dhis2.data.forms.dataentry.tablefields.FormViewHolder;
 import org.dhis2.data.forms.dataentry.tablefields.RowAction;
 import org.dhis2.databinding.CustomCellViewBinding;
+import org.dhis2.usescases.coodinates.CoordinatesView;
 import org.dhis2.utils.DialogClickListener;
-import org.dhis2.utils.customviews.CoordinatesView;
 import org.dhis2.utils.customviews.TableFieldDialog;
 import org.hisp.dhis.android.core.common.FeatureType;
