+++ app/src/main/res/layout-land/activity_event_initial.xml
@@ -237,7 +237,7 @@
                             android:layout_height="wrap_content"
                             android:orientation="vertical" />
 
-                        <org.dhis2.utils.customviews.CoordinatesView
+                        <org.dhis2.usescases.coodinates.CoordinatesView
                             android:id="@+id/geometry"
                             android:layout_width="match_parent"
                             android:layout_height="wrap_content"
